Pull out common test view and proxy element
