+++ Example/Unit Tests/AnimationSnapshotTests.swift
@@ -22,9 +22,9 @@ final class AnimationSnapshotTests: SnapshotTestCase {
     // MARK: - Tests - Frame Snapshots
 
     func testSimpleAnimationSnapshot() {
-        let view = View(frame: .init(x: 0, y: 0, width: 200, height: 40))
+        let view = AnimatableContainerView(frame: .init(x: 0, y: 0, width: 200, height: 40))
 
-        var animation = Animation<View>()
+        var animation = Animation<AnimatableContainerView>()
         animation.addKeyframe(for: \.animatableView.transform, at: 0, value: .identity)
         animation.addKeyframe(for: \.animatableView.transform, at: 1, value: .init(translationX: 160, y: 0))
 
@@ -38,11 +38,11 @@ final class AnimationSnapshotTests: SnapshotTestCase {
     }
 
     func testAnimationWithNonViewElementSnapshot() {
-        let view = View(frame: .init(x: 0, y: 0, width: 200, height: 40))
+        let view = AnimatableContainerView(frame: .init(x: 0, y: 0, width: 200, height: 40))
 
-        let element = Proxy(view: view)
+        let element = AnimatableContainerView.Proxy(view: view)
 
-        var animation = Animation<Proxy>()
+        var animation = Animation<AnimatableContainerView.Proxy>()
         animation.addKeyframe(for: \.animatableViewTransform, at: 0, value: .identity)
         animation.addKeyframe(for: \.animatableViewTransform, at: 1, value: .init(translationX: 160, y: 0))
 
@@ -56,9 +56,9 @@ final class AnimationSnapshotTests: SnapshotTestCase {
     }
 
     func testAnimationWithExecutionBlocksSnapshot() {
-        let view = View(frame: .init(x: 0, y: 0, width: 200, height: 40))
+        let view = AnimatableContainerView(frame: .init(x: 0, y: 0, width: 200, height: 40))
 
-        var animation = Animation<View>()
+        var animation = Animation<AnimatableContainerView>()
         animation.addKeyframe(for: \.animatableView.transform, at: 0, value: .identity)
         animation.addKeyframe(for: \.animatableView.transform, at: 1, value: .init(translationX: 160, y: 0))
         animation.addExecution(
@@ -79,9 +79,9 @@ final class AnimationSnapshotTests: SnapshotTestCase {
     // MARK: - Tests - Animated PNG
 
     func testSimpleAnimationSnapshotAPNG() {
-        let view = View(frame: .init(x: 0, y: 0, width: 200, height: 40))
+        let view = AnimatableContainerView(frame: .init(x: 0, y: 0, width: 200, height: 40))
 
-        var animation = Animation<View>()
+        var animation = Animation<AnimatableContainerView>()
         animation.addKeyframe(for: \.animatableView.transform, at: 0, value: .identity)
         animation.addKeyframe(for: \.animatableView.transform, at: 1, value: .init(translationX: 160, y: 0))
 
@@ -89,9 +89,9 @@ final class AnimationSnapshotTests: SnapshotTestCase {
     }
 
     func testSimpleAnimationSnapshotAPNGAtHighFPS() {
-        let view = View(frame: .init(x: 0, y: 0, width: 200, height: 40))
+        let view = AnimatableContainerView(frame: .init(x: 0, y: 0, width: 200, height: 40))
 
-        var animation = Animation<View>()
+        var animation = Animation<AnimatableContainerView>()
         animation.addKeyframe(for: \.animatableView.transform, at: 0, value: .identity)
         animation.addKeyframe(for: \.animatableView.transform, at: 1, value: .init(translationX: 160, y: 0))
 
@@ -99,9 +99,9 @@ final class AnimationSnapshotTests: SnapshotTestCase {
     }
 
     func testLongAnimationSnapshotAPNG() {
-        let view = View(frame: .init(x: 0, y: 0, width: 200, height: 40))
+        let view = AnimatableContainerView(frame: .init(x: 0, y: 0, width: 200, height: 40))
 
-        var animation = Animation<View>()
+        var animation = Animation<AnimatableContainerView>()
         animation.addKeyframe(for: \.animatableView.transform, at: 0, value: .identity)
         animation.addKeyframe(for: \.animatableView.transform, at: 1, value: .init(translationX: 160, y: 0))
         animation.implicitDuration = 2
@@ -110,9 +110,9 @@ final class AnimationSnapshotTests: SnapshotTestCase {
     }
 
     func testAutoreversingAnimationSnapshotAPNG() {
-        let view = View(frame: .init(x: 0, y: 0, width: 200, height: 40))
+        let view = AnimatableContainerView(frame: .init(x: 0, y: 0, width: 200, height: 40))
 
-        var animation = Animation<View>()
+        var animation = Animation<AnimatableContainerView>()
         animation.addKeyframe(for: \.animatableView.transform, at: 0, value: .identity)
         animation.addKeyframe(for: \.animatableView.transform, at: 1, value: .init(translationX: 160, y: 0))
         animation.implicitRepeatStyle = .infinitelyRepeating(autoreversing: true)
@@ -121,9 +121,9 @@ final class AnimationSnapshotTests: SnapshotTestCase {
     }
 
     func testAnimationWithExecutionBlocksSnapshotAPNG() {
-        let view = View(frame: .init(x: 0, y: 0, width: 200, height: 40))
+        let view = AnimatableContainerView(frame: .init(x: 0, y: 0, width: 200, height: 40))
 
-        var animation = Animation<View>()
+        var animation = Animation<AnimatableContainerView>()
         animation.addKeyframe(for: \.animatableView.transform, at: 0, value: .identity)
         animation.addKeyframe(for: \.animatableView.transform, at: 1, value: .init(translationX: 160, y: 0))
         animation.addExecution(
@@ -136,9 +136,9 @@ final class AnimationSnapshotTests: SnapshotTestCase {
     }
 
     func testAutoreversingAnimationWithExecutionBlocksSnapshotAPNG() {
-        let view = View(frame: .init(x: 0, y: 0, width: 200, height: 40))
+        let view = AnimatableContainerView(frame: .init(x: 0, y: 0, width: 200, height: 40))
 
-        var animation = Animation<View>()
+        var animation = Animation<AnimatableContainerView>()
         animation.addKeyframe(for: \.animatableView.transform, at: 0, value: .identity)
         animation.addKeyframe(for: \.animatableView.transform, at: 1, value: .init(translationX: 160, y: 0))
         animation.addExecution(
@@ -152,11 +152,11 @@ final class AnimationSnapshotTests: SnapshotTestCase {
     }
 
     func testAnimationWithNonViewElementSnapshotAPNG() {
-        let view = View(frame: .init(x: 0, y: 0, width: 200, height: 40))
+        let view = AnimatableContainerView(frame: .init(x: 0, y: 0, width: 200, height: 40))
 
-        let element = Proxy(view: view)
+        let element = AnimatableContainerView.Proxy(view: view)
 
-        var animation = Animation<Proxy>()
+        var animation = Animation<AnimatableContainerView.Proxy>()
         animation.addKeyframe(for: \.animatableViewTransform, at: 0, value: .identity)
         animation.addKeyframe(for: \.animatableViewTransform, at: 1, value: .init(translationX: 160, y: 0))
 
@@ -173,72 +173,6 @@ final class AnimationSnapshotTests: SnapshotTestCase {
         }
 
         SnapshotVerify(animation: animation, on: view)
-    }
-
-}
-
-// MARK: -
-
-extension AnimationSnapshotTests {
-
-    final class View: UIView {
-
-        // MARK: - Life Cycle
-
-        override init(frame: CGRect) {
-            super.init(frame: frame)
-
-            animatableView.backgroundColor = .red
-            addSubview(animatableView)
-        }
-
-        @available(*, unavailable)
-        required init?(coder: NSCoder) {
-            fatalError("init(coder:) has not been implemented")
-        }
-
-        // MARK: - Public Properties
-
-        let animatableView: UIView = .init()
-
-        // MARK: - UIView
-
-        override func layoutSubviews() {
-            animatableView.bounds.size = .init(width: 20, height: 20)
-            animatableView.center = .init(x: 20, y: bounds.midY)
-        }
-
-    }
-
-}
-
-// MARK: -
-
-extension AnimationSnapshotTests {
-
-    final class Proxy {
-
-        // MARK: - Life Cycle
-
-        init(view: View) {
-            self.view = view
-        }
-
-        // MARK: - Public Properties
-
-        public var animatableViewTransform: CGAffineTransform {
-            get {
-                return view.animatableView.transform
-            }
-            set {
-                view.animatableView.transform = newValue
-            }
-        }
-
-        // MARK: - Private Properties
-
-        private let view: View
-
     }
 
 }
